docs(PPDSC-2204): format links correctly
